Research quickfix

See merge request o_spiegl/tso500_dragen_pipeline!37

+++ scripts/processing.py
@@ -14,19 +14,19 @@ def create_parser():
     parser.add_argument('-t', '--testing',action='store_true', help='Testing mode')
     parser.add_argument('-tf', '--testing_fast',action='store_true', help='Fast testing mode')
 
-    return parser
+    return parser.parse_args()
 
 
 def main():
-    args = create_parser().parse_args()
-    testing: bool = args.testing
-    testing_fast: bool = args.testing_fast
+    args = create_parser()
+    testing = args.testing
+    testing_fast = args.testing_fast
     repo_root = get_repo_root()
 
     with open(f'{repo_root}/config.yaml', 'r') as file:
         config = yaml.safe_load(file)
-        servers: list = config['available_servers']
-        server_availability_dir: Path = Path(config['server_availability_dir'])
+        servers = config['available_servers']
+        server_availability_dir = Path(config['server_availability_dir'])
         server = get_server_ip()
         idle_tag = server_availability_dir / server / config['server_idle_tag']
         busy_tag = server_availability_dir / server / config['server_busy_tag']
@@ -45,7 +45,7 @@ def main():
     busy_tag.touch()
     idle_tag.unlink()
 
-    path, input_type, _, tag, flowcell = queue.iloc[0]
+    input_path, input_type, _, tag, flowcell = queue.iloc[0]
 
     last_sample_queue = False
     if input_type == 'sample' and len(queue['Tag'][queue['Tag'] == tag]) == 1:
@@ -60,21 +60,22 @@ def main():
     if len(queue_merged['Flowcell'][queue_merged['Flowcell'] == flowcell]) == 0:
         last_sample_run = True
 
-    paths: dict = setup_paths(repo_root=repo_root, input_path=Path(path), input_type=input_type, tag=tag, flowcell=flowcell, config=config,
+    paths = setup_paths(repo_root=repo_root, input_path=Path(input_path), input_type=input_type, tag=tag, flowcell=flowcell, config=config,
                               testing=testing, testing_fast=testing_fast)
-    failed_tag = paths['failed_tag']
-    analyzing_tag = paths['analyzing_tag']
     try:
         logger = setup_logger(logger_name='Logger',log_file=paths['log_file'])
         check_mountpoint(paths=paths, logger=logger)
         check_structure(paths=paths, logger=logger)
-        check_docker_image(logger=logger)
+        check_docker_image(paths=paths, logger=logger)
         check_rsync(paths=paths, logger=logger)
         check_tso500_script(paths=paths, logger=logger)
 
-        if not paths['analyzing_tag'].exists():
+        if paths['failed_tag'].exists() and not paths['analyzing_tag'].exists():
+            paths['analyzing_tag'].touch()
+        elif not paths['failed_tag'].exists() and not paths['analyzing_tag'].exists():
             paths['analyzing_tag'].touch()
             paths['queued_tag'].unlink()
+
         stage_object(paths=paths, input_type=input_type, last_sample_queue=last_sample_queue, logger=logger)
 
         process_object(paths=paths, input_type=input_type, last_sample_queue=last_sample_queue, logger=logger)
@@ -91,8 +92,8 @@ def main():
             paths['analyzing_tag'].unlink()
 
     except Exception:
-        failed_tag.touch()
-        analyzing_tag.unlink()
+        paths['failed_tag'].touch()
+        paths['analyzing_tag'].unlink()
         raise RuntimeError
     finally:
         idle_tag.touch()
